+++ tests/testthat/test-survival_adapt.R
@@ -25,6 +25,38 @@ test_that("survival_adapt-bayes", {
   expect_s3_class(out, "data.frame")
 })
 
+test_that("survival_adapt-bayes-bin", {
+  set.seed(1)
+  out <- survival_adapt(
+    hazard_treatment = -log(0.85) / 36,
+    hazard_control = -log(0.7) / 36,
+    cutpoints = 0,
+    N_total = 400,
+    lambda = 20,
+    lambda_time = 0,
+    interim_look = 200,
+    end_of_study = 36,
+    prior = c(0.1, 0.1),
+    bin_prior = c(1, 1),
+    bin_method = "normal",
+    bin_N = 100,
+    block = 2,
+    rand_ratio = c(1, 1),
+    prop_loss = 0.30,
+    alternative = "less",
+    h0 = 0,
+    Fn = 0.05,
+    Sn = 0.9,
+    prob_ha = 0.975,
+    N_impute = 2,
+    N_mcmc = 2,
+    method = "bayes-bin"
+  )
+
+  expect_s3_class(out, "data.frame")
+  expect_true(!is.na(out$est_final))
+})
+
 test_that("survival_adapt-logrank", {
   out <- survival_adapt(
     hazard_treatment = -log(0.85) / 36,